pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-openai",
    "langchain-mcp-adapters",
    "python-dotenv",
    "streamlit",
    "fastmcp>=3.1.1",
    "langchain-anthropic>=1.4.0",
]
